Fix score search: expect float instead of integer
Search by score is broken after the update of moving to Wilson score. This commit fixes it.

The issue is, that QPixel expects an integer as score (to_i). Hence `score:>0.7` are interpreted as `score:>0`, which is obviously wrong. The fix should be just changing from to_i to to_f.

+++ app/helpers/search_helper.rb
@@ -26,7 +26,7 @@ module SearchHelper
         next unless value.match?(valid_value[:numeric])
 
         operator, val = numeric_value_sql value
-        ["score #{operator.present? ? operator : '='} ?", val.to_i]
+        ["score #{operator.present? ? operator : '='} ?", val.to_f]
       when 'created'
         next unless value.match?(valid_value[:date])
